requirements.txt
attrs==18.1.0
awesome-slugify==1.6.5
Django==2.0.6
Pillow==5.1.0
bleach==2.1.3
beautifulsoup4==4.6.0
feedparser==5.2.1
lxml==4.2.1
pytz==2018.4
pyTelegramBotAPI==3.6.3
raven==6.9.0
regex==2018.02.21
requests==2.18.4
twitter==1.18.0
django-admirarchy==0.2.2
django-datetime-widget==0.9.3
django-debug-toolbar==1.9.1
django-etc==0.10.0
django-simple-history==2.0.0
git+https://github.com/jazzband/django-robots.git@3.1.0
django-siteblocks==0.6.1
django-sitecats==0.7.0
django-siteflags==0.5.0
django-sitegate==0.13.0
django-sitemessage==0.9.1
django-sitemetrics==0.7.0
django-siteprefs==0.7.0
django-sitetree==1.10.0
django-xross==0.6.0
envbox==0.3.0
uwsgiconf[cli]==0.14.1